Allow filtering failing checks task by schedule
In order to send out specific notifications about checks that are
failing we want to allow the checks to be filtered by the schedule they
are set to.

The current implementation allows us to set up tasks in order to send
the failing checks as a summary potentially once a day if you set up a
new cron task for example.

This implementation gives us the ability to set up multiple cron tasks
to send out summaries filtered down to their schedule. So in practice I
could set up a summary email for daily checks/monthly checks/weekly
checks and only get summaries of the checks with those schedules.